[Runtime] Move some of the changes back to int. (#7529)
Some of the fixes done for the warnings have breaking changes. Move back
to int and ensure that we do not have any compilation errors (we are
using -Werror).

Fixes: https://github.com/xamarin/xamarin-macios/issues/7509

+++ runtime/monotouch-debug.h
@@ -23,7 +23,7 @@ void monotouch_start_debugging ();
 void monotouch_start_profiling ();
 
 void monotouch_set_connection_mode (const char *mode);
-void monotouch_set_monodevelop_port (long port);
+void monotouch_set_monodevelop_port (int port);
 
 
 typedef struct {
